cNetwork: Split cNetworkSingleton to a separate file.
