+++ tests/Network/CMakeLists.txt
@@ -6,10 +6,13 @@ include_directories(${CMAKE_SOURCE_DIR}/src/)
 include_directories(${CMAKE_SOURCE_DIR}/lib/libevent/include)
 
 add_definitions(-DTEST_GLOBALS=1)
+
+# Create a single Network library that contains all the networking code:
 add_library(Network
-	${CMAKE_SOURCE_DIR}/src/OSSupport/Network.cpp
 	${CMAKE_SOURCE_DIR}/src/OSSupport/CriticalSection.cpp
 	${CMAKE_SOURCE_DIR}/src/OSSupport/Event.cpp
+	${CMAKE_SOURCE_DIR}/src/OSSupport/Network.cpp
+	${CMAKE_SOURCE_DIR}/src/OSSupport/NetworkSingleton.cpp
 	${CMAKE_SOURCE_DIR}/src/StringUtils.cpp
 )
 
@@ -18,12 +21,20 @@ if (MSVC)
 	target_link_libraries(Network ws2_32.lib)
 endif()
 
+
+
+
+# Define individual tests:
+
+# Google: download the google.com frontpage using http client socket:
 add_executable(Google-exe Google.cpp)
 target_link_libraries(Google-exe Network)
 add_test(NAME Google-test COMMAND Google-exe)
 
+# EchoServer: Listen on port 9876, echo everything back:
 add_executable(EchoServer EchoServer.cpp)
 target_link_libraries(EchoServer Network)
 
+# NameLookup: Lookup hostname-to-IP and IP-to-hostname:
 add_executable(NameLookup NameLookup.cpp)
 target_link_libraries(NameLookup Network)
